requirements.txt
PyPDF2==3.0.1
pdfplumber==0.11.4
reportlab==4.2.5
fillpdf==0.7.3
pdf-annotate==0.12.0
click==8.1.7
rich==13.9.4
python-dotenv==1.0.1
PyYAML==6.0.1
pytest==8.0.0
pytest-cov==4.1.0